Add handling of retcodes for runner returns
Adds handling of retcodes for runner returns and adds
--retcode-passthrough option.

If --retcode-passthrough option is set for local, local_* or ssh
clients (ie those that return a list of minions) in results, it will
just act as if --fail-any is passed.

If --retcode-passthrough is set for other clients runner*, wheel* etc,
then it will check for a nested return with a retcode. Note it will only
check for a retcode at in [{'data':{'return': {'retcode': <val>}}}]
level. That should be fine as the runner should work out the correct
retcode value.